contentsフォルダをcontentに名称変更
