+++ src/app/content/new/page.tsx
@@ -1,0 +1,12 @@
+import ContentLayout from "@/features/content/components/ContentLayout";
+
+export default function NewContentsPage() {
+  return (
+    <ContentLayout>
+      <div>
+        <h1 className="text-2xl font-bold mb-6">新規コンテンツ作成</h1>
+        {/* 新規コンテンツ作成フォームなど */}
+      </div>
+    </ContentLayout>
+  );
+}
