--- src/app/contents/new/page.tsx
@@ -1,7 +1,0 @@
-export default function NewContentsPage() {
-  return (
-    <div className="container mx-auto px-4 py-8">
-      <h1 className="text-2xl font-bold mb-6">新規コンテンツ作成</h1>
-    </div>
-  );
-}
